requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.0.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.11
dj-database-url==2.1.0

